fix: correct pulse width format in echo line output
Update the printf format specifier for pulse_width_ns from
"%llu" to "%lu" to ensure proper output for the pulse width
in the echo line event handling.

+++ main.c
@@ -353,7 +353,7 @@ int main(int argc, char *argv[]) {
             if (redis_host()) {
               OCCURS(echo, pulse_width_ns);
             } else {
-              (void)printf("%llu\n", (unsigned long long)pulse_width_ns);
+              (void)printf("%lu\n", (unsigned long)pulse_width_ns);
             }
           } else {
             pr_warn("Received falling edge event on echo line with zero pulse width\n");
